refactor(database): emit UserPasskey SETs without coalesce
Drop the patch/coalesce helpers so each typed update writes one SET.
Use uint32 for increment Delta so negatives do not compile and the
value always fits in signed BIGINT/INT64.

+++ internal/domain/user_passkey_update.go
@@ -27,7 +27,7 @@ type UserPasskeySignCountUpdate struct {
 func (*UserPasskeySignCountUpdate) userPasskeyUpdate() {}
 
 type UserPasskeyIncrementSignCountUpdate struct {
-	Delta int64
+	Delta uint32
 }
 
 func (*UserPasskeyIncrementSignCountUpdate) userPasskeyUpdate() {}
